Project-wide code cleanup.

+++ common/src/main/java/com/forerunnergames/tools/common/Utils.java
@@ -35,7 +35,7 @@ public final class Utils
 
     assert foundKey;
 
-    return boxedKey.intValue ();
+    return boxedKey;
   }
 
   /**
